Inject new entries through config instead of custom plugin
