+++ packages/next/package.json
@@ -106,8 +106,9 @@
     "use-subscription": "1.4.1",
     "watchpack": "2.0.0-beta.13",
     "web-vitals": "0.2.1",
-    "webpack": "4.43.0",
-    "webpack-sources": "1.4.3"
+    "webpack": "5.0.0-beta.18",
+    "webpack-sources": "1.4.3",
+    "webpack-dev-middleware": "3.7.2"
   },
   "peerDependencies": {
     "react": "^16.6.0",
@@ -213,7 +214,6 @@
     "typescript": "3.8.3",
     "unfetch": "4.1.0",
     "unistore": "3.4.1",
-    "webpack-dev-middleware": "3.7.0",
     "webpack-hot-middleware": "2.25.0"
   },
   "engines": {
