+++ packages/next/types/misc.d.ts
@@ -210,10 +210,10 @@ declare module 'next/dist/compiled/unistore' {
   import m from 'unistore'
   export = m
 }
-declare module 'next/dist/compiled/webpack-dev-middleware' {
-  import m from 'webpack-dev-middleware'
-  export = m
-}
+// declare module 'next/dist/compiled/webpack-dev-middleware' {
+//   import m from 'webpack-dev-middleware'
+//   export = m
+// }
 declare module 'next/dist/compiled/webpack-hot-middleware' {
   import m from 'webpack-hot-middleware'
   export = m
@@ -248,6 +248,7 @@ declare module 'watchpack' {
   import { EventEmitter } from 'events'
 
   class Watchpack extends EventEmitter {
+    constructor(options?: any)
     watch(files: string[], directories: string[], startTime?: number): void
     close(): void
